test(ui): add UploadZone file validation tests
